fixing warnings and code style, refs #3
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@9477 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/devices/MSDevice_Vehroutes.cpp
@@ -158,7 +158,7 @@ MSDevice_Vehroutes::writeXMLRoute(OutputDevice &os, int index) const {
         unsigned int edgeIndex = 0;
         if (myHolder.getNumberReroutes() > 0) {
             assert(myReplacedRoutes.size()<=myHolder.getNumberReroutes());
-            unsigned int i = myReplacedRoutes.size();
+            unsigned int i = static_cast<unsigned int>(myReplacedRoutes.size());
             while (i > 0 && myReplacedRoutes[i-1].edge) {
                 i--;
             }
